Add tag and story gateway
