+++ apps/gateway/src/modules/tag/tag-gateway.controller.ts
@@ -9,10 +9,7 @@ import { TagEntity } from "../../../../tag/src/entity/tag.entity";
 import { IGatewayResponse } from "../../common/interface/gateway.interface";
 import { Auth } from "@app/authentication";
 import { TAG_MESSAGE_PATTERNS } from "../../../../tag/src/constant/tag-patterns.constants";
-import { FindMemberDto } from "../../../../member/src/dto/find.member.dto";
 import { IPagination } from "@app/common";
-import { MemberEntity } from "../../../../member/src/entity/member.entity";
-import { MEMBER_MESSAGE_PATTERNS } from "../../../../member/src/constant/member-patterns.constants";
 import { AllTagsAppDto } from "../../../../tag/src/dto/all.tags.app.dto";
 
 @ApiTags('Tags Gateway')
@@ -23,14 +20,12 @@ export class TagGatewayController {
 
   constructor(@Inject(RabbitServiceName.TAG) private tagClient: ClientProxy) { }
 
-
-  @Auth()
-  @Get('/all')
+  @Get('/app')
   async getTagsApp ( @Query() findDto: AllTagsAppDto ): Promise<IGatewayResponse<IPagination<TagEntity>>> {
     const { state, data } = await firstValueFrom(
       this.tagClient.send<IServiceResponse<IPagination<TagEntity>>>
       (
-        TAG_MESSAGE_PATTERNS.ALL,
+        TAG_MESSAGE_PATTERNS.APP,
         findDto
       )
     );
@@ -42,8 +37,7 @@ export class TagGatewayController {
   @Auth()
   @ApiOperation({ summary: 'Tag Story' })
   @ApiResponse({ status: 201, description: 'Add a Tag story' })
-  async create (
-    @Body() tagDto: CreateTagDto,) : Promise<IGatewayResponse> {
+  async create (@Body() tagDto: CreateTagDto,) : Promise<IGatewayResponse> {
 
     const { state, data, message } = await firstValueFrom(
       this.tagClient.send<IServiceResponse<TagEntity>,
